Adding a plugins folder
