+++ airflow/dags/karpenter_test.py
@@ -1,8 +1,9 @@
 from datetime import datetime
 
-from airflow.providers.cncf.kubernetes.operators.pod import KubernetesPodOperator
+from unity_sps_utils import get_affinity
 
 from airflow import DAG
+from airflow.providers.cncf.kubernetes.operators.pod import KubernetesPodOperator
 
 POD_NAMESPACE = "airflow"
 POD_LABEL = "karpenter_test_task"
@@ -13,65 +14,6 @@ default_args = {
 }
 
 default_params = {"placeholder": 1}
-
-
-def get_affinity(
-    capacity_type: list[str], instance_family=list[str], instance_cpu=list[str], anti_affinity_label=str
-):
-
-    affinity = {
-        "nodeAffinity": {
-            "preferredDuringSchedulingIgnoredDuringExecution": [
-                {
-                    "weight": 1,
-                    "preference": {
-                        "matchExpressions": [
-                            {
-                                "key": "karpenter.sh/capacity-type",
-                                "operator": "In",
-                                "values": capacity_type,
-                            }
-                        ]
-                    },
-                }
-            ],
-            "requiredDuringSchedulingIgnoredDuringExecution": {
-                "nodeSelectorTerms": [
-                    {
-                        "matchExpressions": [
-                            {
-                                "key": "karpenter.k8s.aws/instance-family",
-                                "operator": "In",
-                                "values": instance_family,
-                            },
-                            {
-                                "key": "karpenter.k8s.aws/instance-cpu",
-                                "operator": "In",
-                                "values": instance_cpu,
-                            },
-                        ]
-                    }
-                ]
-            },
-        },
-        "podAntiAffinity": {
-            "requiredDuringSchedulingIgnoredDuringExecution": [
-                {
-                    "labelSelector": {
-                        "matchExpressions": [
-                            {
-                                "key": "app",
-                                "operator": "In",
-                                "values": [anti_affinity_label],
-                            },
-                        ]
-                    },
-                    "topologyKey": "kubernetes.io/hostname",
-                }
-            ]
-        },
-    }
-    return affinity
 
 
 dag = DAG(
