+++ airflow/docker/custom_airflow/Dockerfile
@@ -2,3 +2,5 @@ FROM apache/airflow:2.9.1-python3.11
 
 RUN pip install cwltool==3.1.20240112164112
 RUN pip install boto3==1.34.89
+
+COPY ../plugins /opt/airflow/plugins
